Execute a command

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -2,4 +2,5 @@
 
 exec { 'killmenow_process':
   command  => 'pkill killmenow',
+  onlyif   => 'pgrep killmenow',
 }
